Add oneLineStrip tag to concatenate multiline strings without spaces
This adds a new `oneLineStrip` option and a corresponding tag to allow
multiline strings to be concatenated without any white spaces.`oneLine`
replaces the new lines & indents with a single white space.  However, it
is sometimes needed, e.g. for things like URLs, to not add any white
space but simply trim the new lines & indents.